Revert "[INTERNAL] Move resource content into separate instance"
No major benefits since streams need to be drained into a buffer
anyways.

This reverts commit ecd7bd600c8735ffd946ebadd5bd2a82734a4168.
